warning
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@20401 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ msctl/msctl.c
@@ -245,7 +245,7 @@ parse_replrq(int opcode, const char *fn, char *replrqspec,
 	for (ios = iosv; ios; ios = next) {
 		if ((next = strchr(ios, ',')) != NULL)
 			*next++ = '\0';
-		if (ra.nios >= nitems(ra.iosv))
+		if (ra.nios >= (int)nitems(ra.iosv))
 			errx(1, "%s: too many replicas specified",
 			    replrqspec);
 		if (strchr(ios, '@') == NULL)
